+++ src/Aks/Aks/ChangeLog.md
@@ -19,6 +19,11 @@
 -->
 ## Upcoming Release
 * Bumped API version to 2022-09-01
+* Added parameter `NodeOsSKU` for `New-AzAksCluster` and parameter `OsSKU` for `New-AzAksNodePool`
+* Added parameter `Mode` for `New-AzAksNodePool` and `Update-AzAksNodePool`
+
+## Version 5.0.1
+* Upgraded AutoMapper to Microsoft.Azure.PowerShell.AutoMapper 6.2.2 with fix [#18721]
 
 ## Version 5.0.0
 * [Breaking Change] Removed the alias `Install-AzAksKubectl` of `Install-AzAksCliTool`.